control-api/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
python-jose==3.3.0
pydantic==1.10.14
alembic==1.13.1
python-dotenv==1.0.1
boto3==1.34.34
